release: Chromium 5.0.375 (5.0.375.99)
Compare: https://github.com/chromium/chromium/compare/08be0661f03fc0f45ef2e9bd14244777e9bbe34f...2544c4012ca37e2672b0fb656838d1fb0d76a6d4

+++ o3d/core/cross/gl/effect_gl.cc
@@ -41,7 +41,6 @@
 #pragma warning(disable : 4311)
 #endif
 
-#include "Cg/cgGL.h"
 #include <sstream>
 #include "base/cross/std_functional.h"
 #include "core/cross/semantic_manager.h"
@@ -562,7 +561,8 @@ void EffectGL::SetTexturesFromEffect(ParamCacheGL* param_cache_gl) {
     if (param != NULL) {
       Texture *t = param->value();
       if (t) {
-        GLuint handle = reinterpret_cast<GLuint>(t->GetTextureHandle());
+        GLuint handle = static_cast<GLuint>(reinterpret_cast<intptr_t>(
+            t->GetTextureHandle()));
         cgGLSetTextureParameter(cg_param, handle);
         cgGLEnableTextureParameter(cg_param);
       }
